% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addConstrain.R
\name{addBranch}
\alias{addBranch}
\title{Add a branch to a data.table}
\usage{
addBranch(domain, ptdf, ram, name = "CustomBranch")
}
\arguments{
\item{domain}{{data.frame | data.table}}

\item{ptdf}{{list} ptdf of new contrain, named list}

\item{ram}{{numeric} ram  of new contrain}

\item{name}{{character} name of new contrain.}
}
\description{
Add a branch to a data.table
}
\examples{
\dontrun{
PLAN2 <- addBranch(domain,
ptdf = list(AT = 0, BE = 1, DE = 0, FR = 0, NL = 0),
ram = -1000 )
}

}
