% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addConstrain.R
\name{addConstrain}
\alias{addConstrain}
\title{Add a PTDF to a data.table}
\usage{
addConstrain(PTDF, ctr, ram, name = "CustomConstrain")
}
\arguments{
\item{PTDF}{{data.frame | data.table}}

\item{ctr}{{list} ptdf of new contrain, named list}

\item{ram}{{numeric} ram  of new contrain}

\item{name}{{character} name of new contrain.}
}
\description{
Add a PTDF to a data.table
}
\examples{
\dontrun{
PLAN2 <- addConstrain(PLAN, ctr = list(AT = 0, BE = 1, DE = 0, FR = 0, NL = 0),
ram = -1000 )
}

}
