% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computePn.R
\name{computePn}
\alias{computePn}
\title{Extract ptdf from cnes XML}
\usage{
computePn(LINK, ctry, name = "")
}
\arguments{
\item{LINK}{{data.table} LINK data.table, first column is timestamp others are area1-area2}

\item{ctry}{{character} areas for compute net position.}

\item{name}{{character} prefix for area.}
}
\description{
Extract ptdf from cnes XML
}
\examples{
\dontrun{

   refProf = getRefProgs("D:/Users/titorobe/Desktop/CNESdata/Refprogs/22XCORESO------S_17XTSO-CS------W_CWE-FB-D2CF-100_20180903F10101.xml")
   computePn(refProf, c("BE", "NL", "DE", "FR", "APG"), "CWEP")
   
}

}
