% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPlan.R
\name{distToBranch}
\alias{distToBranch}
\title{extractPnCnes from net position data}
\usage{
distToBranch(PTDF, pos, threshold = NULL)
}
\arguments{
\item{PTDF}{{data.table} PTDF}

\item{pos}{{list} eq : pos = list(AT = 0, BE = 0, DE = 0, FR = 0, NL = 0)}

\item{threshold}{{numeric}, threshold.}
}
\description{
extractPnCnes from net position data
}
\examples{
\dontrun{
   distToBranch(PTDF, pos = list(AT = 0, BE = 0, DE = 0, FR = 0, NL = 0))
}

}
