% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCnesFb.R
\name{getCnesFb}
\alias{getCnesFb}
\title{Extract ptdf from cnes XML}
\usage{
getCnesFb(file, nbcl, sizechunk = 10)
}
\arguments{
\item{file}{{character} XML file}

\item{nbcl}{{numeric} parralel process}

\item{sizechunk}{{numeric} size of memory chunk.}
}
\description{
Extract ptdf from cnes XML
}
\examples{
\dontrun{
   getCnesFb("fb.XML", 4 )
}

}
