% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLtPos.R
\name{getLtPos}
\alias{getLtPos}
\title{extractPnCnes from net position data}
\usage{
getLtPos(xmlLt)
}
\arguments{
\item{xmlLt}{{character} long time file(s)}
}
\description{
extractPnCnes from net position data
}
