% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRefprogs.R
\name{getRefProgs}
\alias{getRefProgs}
\title{Extract RefProgs from XML}
\usage{
getRefProgs(file)
}
\arguments{
\item{file}{{character} XML file(s)}
}
\description{
Extract RefProgs from XML
}
\examples{
\dontrun{
   getRefProgs("fb.XML")
}

}
