% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVertices.R
\name{getVertices}
\alias{getVertices}
\title{vertices from PTDF data}
\usage{
getVertices(PTDF, ctrdel = "NL")
}
\arguments{
\item{PTDF}{{data.frame | data.table}}

\item{ctrdel}{{character} name of country deleted}
}
\description{
vertices from PTDF data
}
\examples{
\dontrun{
  PTDFv <- getVertices(PTDF, "NL")
}
}
