% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{hullPlot}
\alias{hullPlot}
\title{3D flowbased plot}
\usage{
hullPlot(df, grouping, country = c("DE", "BE", "FR"), legend = TRUE,
  alpha = 1)
}
\arguments{
\item{df}{{data.frame} vertices}

\item{grouping}{{character} column for dates}

\item{country}{{character} 3 country (for dims)}

\item{legend}{{boolean} add legend}

\item{alpha}{{numeric} transparancy}
}
\description{
This function generates 3D flowbased plot
}
