% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalOrigin.R
\name{rescalOrigin}
\alias{rescalOrigin}
\title{Rescale data}
\usage{
rescalOrigin(PTDF, PT)
}
\arguments{
\item{PTDF}{{data.frame | data.table}}

\item{PT}{{data.table} Point for rescale.

\dontrun{

 PT = data.table()
 PT$AT <- PT$DE <- PT$FR <- PT$NL <- 0
 PT$BE <- 1
 PT$timestamp  <- '2018-03-03 15:00:00'
 PROJ <- rescalOrigin(PTDF, PT)

}}
}
\description{
Rescale data
}
