% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescalOrigin.R
\name{shiftOriginDomain}
\alias{shiftOriginDomain}
\title{Rescale data}
\usage{
shiftOriginDomain(domain, point)
}
\arguments{
\item{domain}{{data.frame | data.table}}

\item{point}{{data.table} Point for rescale.}
}
\description{
Rescale data
}
\examples{
\dontrun{
 PT = data.table()
 PT$NP_AT <- PT$NP_DE <- PT$NP_FR <- PT$NP_NL <- 0
 PT$NP_BE <- 1
 PT$timestamp  <- '2018-03-03 15:00:00'
 PROJ <- shiftOriginDomain(domain, PT)
}


}
