% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionOnDir.R
\name{getProjPT}
\alias{getProjPT}
\title{Project on direction}
\usage{
getProjPT(PT, PTDF, step = 5)
}
\arguments{
\item{PT}{{data.table} Point for direction to project}

\item{PTDF}{{data.frame | data.table}}

\item{step}{{numeric} manimum step (MW)}
}
\description{
Project on direction
}
\examples{

\dontrun{

 CP = data.table()
 CP$AT <- CP$DE <- CP$FR <- CP$NL <- 0
 CP$BE <- 1
 PROJ <- getProjPT(CP, PTDF)

}

}
