% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectionOnDir.R
\name{projectionOnDir}
\alias{projectionOnDir}
\title{Project on direction}
\usage{
projectionOnDir(domain, point, step = 5)
}
\arguments{
\item{domain}{{data.frame | data.table}}

\item{point}{{data.table} Point for direction to project}

\item{step}{{numeric} manimum step (MW)}
}
\description{
Project on direction
}
\examples{

\dontrun{

 CP = data.table()
 CP$AT <- CP$DE <- CP$FR <- CP$NL <- 0
 CP$BE <- 1
 PROJ <- projectionOnDir(CP, domain)

}

}
